Move slots from result_pane to mainwin
