+++ blackboard/gui/mainwin.py
@@ -2,7 +2,7 @@ from .result_pane import result_pane
 import gi
 gi.require_version('Gtk', '3.0')
 
-from gi.repository import Gtk, Gio  # noqa
+from gi.repository import Gtk, Gio, Gdk  # noqa
 
 
 class mainwin(Gtk.Window):
@@ -16,6 +16,20 @@ class mainwin(Gtk.Window):
         self.__add_paned()
         self.__add_header_bar()
         self.txt_op.grab_focus()
+
+    def __get_cursor_position(self):
+        x = self.txt_op.get_cursor_locations()
+        height = x.strong.height
+        pos = x.strong.y
+        index = int(pos / height)
+        return index
+
+    def add_label_at_cursor(self, txt_op, value):
+        key = Gdk.keyval_name(value.keyval)
+        if key != 'Return':
+            return
+        index = self.__get_cursor_position()
+        self.result_pane.add_label_at(index + 1)
 
     def __create_header_btn(self, icon_name):
         icon = Gio.ThemedIcon(name=icon_name)
@@ -45,7 +59,7 @@ class mainwin(Gtk.Window):
         self.txt_op = Gtk.TextView()
         self.result_pane = result_pane()
         self.txt_op.connect("key-press-event",
-                            self.result_pane.add_label_at_cursor)
+                            self.add_label_at_cursor)
 
     def __add_scrolled_win(self):
         self.scroll = Gtk.ScrolledWindow()
