+++ blackboard/gui/result_pane.py
@@ -1,36 +1,23 @@
 import gi
 gi.require_version('Gtk', '3.0')
 
-from gi.repository import Gtk, Gio, Gdk  # noqa
+from gi.repository import Gtk, Gio  # noqa
 
 
 class result_pane(Gtk.Grid):
     def __init__(self):
         super().__init__()
         self.lb_list = []
-        self.__add_label_at(0)
+        self.add_label_at(0)
 
-    def __add_label_at(self, index):
+    def add_label_at(self, index):
         lb = Gtk.Label()
         self.lb_list.insert(index, lb)
         if (len(self.lb_list) - 1) > index:
             self.insert_row(index)
         self.attach(lb, 0, index, 1, 1)
+        lb.set_label(str(index))
         lb.show()
-
-    def __get_cursor_position(self, txt_op):
-        x = txt_op.get_cursor_locations()
-        height = x.strong.height
-        pos = x.strong.y
-        index = int(pos / height)
-        return index
-
-    def add_label_at_cursor(self, txt_op, value):
-        key = Gdk.keyval_name(value.keyval)
-        if key != 'Return':
-            return
-        index = self.__get_cursor_position(txt_op)
-        self.__add_label_at(index + 1)
 
     def update_label(self, index, result):
         try:
